requirements.txt
streamlit
google-genai
yt-dlp
feedparser
requests
pandas